+++ changelog.md
@@ -6,6 +6,7 @@ This project uses Semantic Versioning (2.0).
 * Add web API.post service to post form data to a url.
 * Update web survey-form to use ursl to get and pot the form data.
 * Move user_id from the api serializer to the url.
+* Add create to the get latest endpoint.
 
 ### 0.0.4
 * Add date_created to the output of latest_response() and its helpers.